app/requirements.txt
trino
scipy
numpy<2
requests==2.31.0
pandas==1.5.3
scikit-learn==1.0
streamlit==1.28.0
altair<5
Pillow
boto3
